update router & create dynamic index menu
