+++ src/common/services/HTTPService.ts
@@ -13,8 +13,6 @@ export type RequestConfig = {
 export default class HTTPService {
     protected instance: AxiosInstance;
 
-    protected instanceRaw: AxiosInstance;
-
     constructor() {
         axios.defaults.withCredentials = true;
 
@@ -41,11 +39,6 @@ export default class HTTPService {
             }
 
             return resp;
-        });
-
-        this.instanceRaw = axios.create({
-            baseURL: '',
-            withCredentials: true
         });
     }
 
@@ -95,16 +88,18 @@ export default class HTTPService {
     }
 
     headRaw(config: Omit<RequestConfig, 'payload'>) {
-        return this.instanceRaw({
+        return this.instance({
             method: 'head',
+            baseURL: '',
             url: config.url,
             signal: config.signal
         });
     }
 
     rawGet(config: RequestConfig) {
-        return this.instanceRaw({
+        return this.instance({
             method: 'get',
+            baseURL: '',
             url: config.url,
             params: config.payload,
             signal: config.signal
