+++ src/components/UI/menu/NavMenu.vue
@@ -106,7 +106,6 @@
     import {
         defineComponent, ref
     } from 'vue';
-    import { tryOnBeforeMount } from '@vueuse/core';
     import { storeToRefs } from 'pinia';
     import { useRoute, useRouter } from 'vue-router';
     import type { TNavItem } from '@/store/UI/NavStore';
@@ -136,10 +135,6 @@
             const router = useRouter();
             const route = useRoute();
             const isShowMenu = ref(false);
-
-            tryOnBeforeMount(async () => {
-                await navStore.initNavItems();
-            });
 
             const isRouteExist = (link: TNavItem) => {
                 if (!link.url) {
